Great Goblins
My first game made in pygame!

Its not very good, but I find it quite fun nonetheless!

If you are on Windows: Once you download the zip file, extract it file and run the executable in it. Make sure that the "Assets" folder in in the same file as the executable.

If you are on Mac/Linux: I have no idea how to export to these platforms. If you really want to play this for some reason, the easiest way is to download the source code and install python/pygame. Then you should be able to run the program from the terminal or from an IDE like vs code. You could also use a tool like virtual box.

I'll might try to make it work on mac/linux in the future, ¯\_(ツ)_/¯